Created Save/Load system, Need to fix horizontal scroll grabber texture

+++ Scripts/Global.gd
@@ -2,7 +2,7 @@ extends Node
 
 var pauseDialogue
 var dict = []
-var happy_expression = preload("res://Art/Sprites/expression_bestie_nervous.png")
+var happy_expression = "res://Art/Sprites/expression_bestie_nervous.png"
 
 var has_signaled = false
 
@@ -14,15 +14,26 @@ func readJSON(var filename: String):
 		return
 	
 	file.open(filename, File.READ)
-	var data=  parse_json(file.get_as_text())
+	var data = parse_json(file.get_as_text())
 	file.close()
 	return data
 
-func writeJSON(var info: String):
+func writeHistory(var info: String):
 	var path = "res://SavedData/history.json"
 	var file = File.new()
-	file.open(path, File.READ_WRITE)
-	file.seek_end()
-	file.store_line(to_json(info))
+	if !file.file_exists(path):
+		file.open(path, File.WRITE)
+		file.store_string(to_json([info]))
+	else:
+		file.open(path, File.READ_WRITE)
+		file.seek_end(-1)
+		file.store_string(",")
+		file.store_line(to_json(info))
+		file.store_string("]")
 	file.close()
 
+func writeJSON(var info: String, var path: String = "res//SavedData/dump.txt"):
+	var file = File.new()
+	file.open(path, File.WRITE)
+	file.store_string(to_json(info))
+	file.close()
